// client/src/api/model/actionType.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * The category of action being taken during structured combat
 */
export type ActionType = typeof ActionType[keyof typeof ActionType];


export const ActionType = {
  Combat_Check: 'Combat Check',
  Skill_Check: 'Skill Check',
  Cast_a_Spell: 'Cast a Spell',
  Use_a_Talent: 'Use a Talent',
  Use_an_ability: 'Use an ability',
  Perform_a_Second_Maneuver: 'Perform a Second Maneuver',
} as const;